Feature/car imges editing

+++ car_auction_service/cars/views.py
@@ -53,11 +53,11 @@ def car_detail(request, pk):
     """
     car = get_object_or_404(Car, pk=pk)
     car_images = reversed(get_list_or_404(CarImage, car=car))
+
     # Adding map component
     # Getting location from car model
     get_car_location = car.location
     location_values = geocoder.osm(f'{get_car_location}, Poland')
-    print(f'-------- {location_values.latlng}')
     # Creating Map Object
     cars_map = folium.Map(location=location_values.latlng, zoom_start=8)
     # Adding map marker
@@ -98,14 +98,7 @@ def delete_car_image(request, car_id, image_id):
     """
     car = get_object_or_404(Car, pk=car_id)
     car_image = get_object_or_404(CarImage, pk=image_id, car=car)
-    print("----------------------------- car")
-    print(car)
-    print("------------------------------- car id")
-    print(car.pk)
-    print('------------------------car image')
-    print(car_image)
-    print('-----------------------')
-    print(car_image.pk)
+
     if request.method == 'POST':
         car_image.delete()
     if len(car.get_all_images()) == 0:
@@ -138,6 +131,7 @@ def car_observe(request, pk):
         messages.success(request, 'Car removed from observed')
 
     return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+
 
 # rozbić dashboard
 @login_required(login_url='/users/accounts/login')
@@ -234,12 +228,16 @@ def car_edit(request, pk):
             car_instance.owner = request.user
             car_instance.save()
 
-            # deleting default car image:
-            if car_image_default is not None:
+            # deleting default car image
+            images_empty = len(request.FILES.getlist('image'))
+            # if there is car image object and it's image is default and there are added images through form
+            # deleting default image only when images added are added through form
+            if car_image_default is not None and car_image_default.image == 'images/no_car_image.png' and images_empty != 0:
                 car_image_default.delete()
 
             # create car images as being related to car object
             images = request.FILES.getlist('image')
+
             for car_image in images:
                 CarImage.objects.create(car=car_instance, image=car_image)
             messages.add_message(request, messages.INFO, 'Car modified')
